rename bot to autonomous task
